+++ test/base/artifact/test_geotiff.py
@@ -25,9 +25,10 @@ def test_create_concise_geotiff_artifact_2d(default_computation_resources, gener
     expected_artifact = _Artifact(
         name='Test Raster',
         modality=ArtifactModality.MAP_LAYER_GEOTIFF,
-        file_path=Path(default_computation_resources.computation_dir / f'{general_uuid}.tiff'),
+        filename='test_file.tiff',
         summary='Raster caption',
         attachments=Attachments(legend=Legend(legend_data={'1': Color('#0f0')})),
+        correlation_uuid=general_uuid,
     )
 
     method_input = RasterInfo(
@@ -49,12 +50,14 @@ def test_create_concise_geotiff_artifact_2d(default_computation_resources, gener
         layer_name='Test Raster',
         caption='Raster caption',
         resources=default_computation_resources,
-        filename=str(general_uuid),
+        filename='test_file',
     )
 
     assert generated_artifact == expected_artifact
 
-    with rasterio.open(generated_artifact.file_path) as generated_content:
+    with rasterio.open(
+        default_computation_resources.computation_dir / generated_artifact.filename
+    ) as generated_content:
         assert_array_equal(generated_content.read(), [method_input.data])
 
 
@@ -64,13 +67,14 @@ def test_create_extensive_geotiff_artifact_2d(
     expected_artifact = _Artifact(
         name='Test Raster',
         modality=ArtifactModality.MAP_LAYER_GEOTIFF,
-        file_path=Path(default_computation_resources.computation_dir / f'{general_uuid}.tiff'),
+        filename='test_file.tiff',
         summary='Raster caption',
         description='Raster description',
         sources=default_sources,
         tags=default_association_tags,
         primary=False,
         attachments=Attachments(legend=Legend(legend_data={'1': Color('#0f0')}, title='Custom Legend Title')),
+        correlation_uuid=general_uuid,
     )
 
     method_input = RasterInfo(
@@ -99,7 +103,7 @@ def test_create_extensive_geotiff_artifact_2d(
         description='Raster description',
         sources=Path(__file__).parent.parent.parent / 'resources/minimal.bib',
         tags=default_association_tags,
-        filename=str(general_uuid),
+        filename='test_file',
     )
 
     assert generated_artifact == expected_artifact
@@ -110,7 +114,9 @@ def test_create_extensive_geotiff_artifact_2d(
     assert method_input.transformation == method_input_copy.transformation
     assert method_input.colormap == method_input_copy.colormap
 
-    with rasterio.open(generated_artifact.file_path) as generated_content:
+    with rasterio.open(
+        default_computation_resources.computation_dir / generated_artifact.filename
+    ) as generated_content:
         assert_array_equal(generated_content.read(), [method_input.data])
 
 
@@ -118,9 +124,10 @@ def test_create_geotiff_artifact_2d_rgba(default_computation_resources, general_
     expected_artifact = _Artifact(
         name='Test Raster',
         modality=ArtifactModality.MAP_LAYER_GEOTIFF,
-        file_path=Path(default_computation_resources.computation_dir / f'{general_uuid}.tiff'),
+        filename='test_file.tiff',
         summary='Raster caption',
         attachments=Attachments(legend=Legend(legend_data={'1': Color('#00ff00fe')})),
+        correlation_uuid=general_uuid,
     )
 
     method_input = RasterInfo(
@@ -142,16 +149,18 @@ def test_create_geotiff_artifact_2d_rgba(default_computation_resources, general_
         layer_name='Test Raster',
         caption='Raster caption',
         resources=default_computation_resources,
-        filename=str(general_uuid),
+        filename='test_file',
     )
 
     assert generated_artifact == expected_artifact
 
-    with rasterio.open(generated_artifact.file_path) as generated_content:
+    with rasterio.open(
+        default_computation_resources.computation_dir / generated_artifact.filename
+    ) as generated_content:
         assert_array_equal(generated_content.read(), [method_input.data])
 
 
-def test_create_geotiff_artifact_masked_array(default_computation_resources, general_uuid):
+def test_create_geotiff_artifact_masked_array(default_computation_resources):
     method_input = RasterInfo(
         data=np.ma.masked_array(np.ones(shape=(2, 2), dtype=np.uint8), mask=[[0, 0], [0, 1]]),
         crs=CRS({'init': 'epsg:4326'}),
@@ -171,10 +180,12 @@ def test_create_geotiff_artifact_masked_array(default_computation_resources, gen
         layer_name='Test Raster',
         caption='Raster caption',
         resources=default_computation_resources,
-        filename=str(general_uuid),
+        filename='test_file',
     )
 
-    with rasterio.open(generated_artifact.file_path) as generated_content:
+    with rasterio.open(
+        default_computation_resources.computation_dir / generated_artifact.filename
+    ) as generated_content:
         assert_array_equal(generated_content.read(1, masked=True), method_input.data)
 
 
@@ -182,9 +193,10 @@ def test_create_geotiff_with_legend_data(default_computation_resources, general_
     expected_artifact = _Artifact(
         name='Test Raster',
         modality=ArtifactModality.MAP_LAYER_GEOTIFF,
-        file_path=Path(default_computation_resources.computation_dir / f'{general_uuid}.tiff'),
+        filename='test_file.tiff',
         summary='Raster caption',
         attachments=Attachments(legend=Legend(legend_data={'A': Color('#f00')})),
+        correlation_uuid=general_uuid,
     )
 
     method_input = RasterInfo(
@@ -206,13 +218,15 @@ def test_create_geotiff_with_legend_data(default_computation_resources, general_
         layer_name='Test Raster',
         caption='Raster caption',
         resources=default_computation_resources,
-        filename=str(general_uuid),
+        filename='test_file',
         legend=Legend(legend_data={'A': Color('red')}),
     )
 
     assert generated_artifact == expected_artifact
 
-    with rasterio.open(generated_artifact.file_path) as generated_content:
+    with rasterio.open(
+        default_computation_resources.computation_dir / generated_artifact.filename
+    ) as generated_content:
         assert_array_equal(generated_content.read(), [method_input.data])
 
 
@@ -220,8 +234,9 @@ def test_create_geotiff_without_legend(default_computation_resources, general_uu
     expected_artifact = _Artifact(
         name='Test Raster',
         modality=ArtifactModality.MAP_LAYER_GEOTIFF,
-        file_path=Path(default_computation_resources.computation_dir / f'{general_uuid}.tiff'),
+        filename='test_file.tiff',
         summary='Raster caption',
+        correlation_uuid=general_uuid,
     )
 
     method_input = RasterInfo(
@@ -242,12 +257,14 @@ def test_create_geotiff_without_legend(default_computation_resources, general_uu
         layer_name='Test Raster',
         caption='Raster caption',
         resources=default_computation_resources,
-        filename=str(general_uuid),
+        filename='test_file',
     )
 
     assert generated_artifact == expected_artifact
 
-    with rasterio.open(generated_artifact.file_path) as generated_content:
+    with rasterio.open(
+        default_computation_resources.computation_dir / generated_artifact.filename
+    ) as generated_content:
         assert_array_equal(generated_content.read(), [method_input.data])
 
 
@@ -255,9 +272,10 @@ def test_create_geotiff_artifact_3d(default_computation_resources, general_uuid)
     expected_artifact = _Artifact(
         name='Test Raster',
         modality=ArtifactModality.MAP_LAYER_GEOTIFF,
-        file_path=Path(default_computation_resources.computation_dir / f'{general_uuid}.tiff'),
+        filename='test_file.tiff',
         summary='Raster caption',
         attachments=Attachments(legend=Legend(legend_data={'1': Color('#0f0')})),
+        correlation_uuid=general_uuid,
     )
 
     method_input = RasterInfo(
@@ -279,16 +297,18 @@ def test_create_geotiff_artifact_3d(default_computation_resources, general_uuid)
         layer_name='Test Raster',
         caption='Raster caption',
         resources=default_computation_resources,
-        filename=str(general_uuid),
+        filename='test_file',
     )
 
     assert generated_artifact == expected_artifact
 
-    with rasterio.open(generated_artifact.file_path) as generated_content:
+    with rasterio.open(
+        default_computation_resources.computation_dir / generated_artifact.filename
+    ) as generated_content:
         assert_array_equal(generated_content.read(), method_input.data)
 
 
-def test_geotiff_creation_option_warnings(default_computation_resources, general_uuid, caplog):
+def test_geotiff_creation_option_warnings(default_computation_resources, caplog):
     method_input = RasterInfo(
         data=np.ones(shape=(4, 5), dtype=np.uint8),
         crs=CRS({'init': 'epsg:4326'}),
@@ -309,13 +329,13 @@ def test_geotiff_creation_option_warnings(default_computation_resources, general
             layer_name='Test Raster',
             caption='Raster caption',
             resources=default_computation_resources,
-            filename=str(general_uuid),
+            filename='test_file',
         )
 
     assert caplog.messages == []
 
 
-def test_small_geotiff_created_without_overviews(default_computation_resources, general_uuid):
+def test_small_geotiff_created_without_overviews(default_computation_resources):
     method_input = RasterInfo(
         data=np.ones(shape=(10, 10), dtype=np.uint8),
         crs=CRS({'init': 'epsg:4326'}),
@@ -334,14 +354,16 @@ def test_small_geotiff_created_without_overviews(default_computation_resources, 
         layer_name='Test Raster',
         caption='Raster caption',
         resources=default_computation_resources,
-        filename=str(general_uuid),
+        filename='test_file',
     )
 
-    with rasterio.open(generated_artifact.file_path) as generated_content:
+    with rasterio.open(
+        default_computation_resources.computation_dir / generated_artifact.filename
+    ) as generated_content:
         assert generated_content.overviews(1) == []
 
 
-def test_geotiff_created_with_overviews(default_computation_resources, general_uuid):
+def test_geotiff_created_with_overviews(default_computation_resources):
     method_input = RasterInfo(
         data=np.ones(shape=(1000, 1000), dtype=np.uint8),
         crs=CRS({'init': 'epsg:4326'}),
@@ -361,14 +383,16 @@ def test_geotiff_created_with_overviews(default_computation_resources, general_u
         layer_name='Test Raster',
         caption='Raster caption',
         resources=default_computation_resources,
-        filename=str(general_uuid),
+        filename='test_file',
     )
 
-    with rasterio.open(generated_artifact.file_path) as generated_content:
+    with rasterio.open(
+        default_computation_resources.computation_dir / generated_artifact.filename
+    ) as generated_content:
         assert generated_content.overviews(1) == [2, 4]
 
 
-def test_geotiff_tiled(default_computation_resources, general_uuid):
+def test_geotiff_tiled(default_computation_resources):
     method_input = RasterInfo(
         data=np.ones(shape=(1000, 1000), dtype=np.uint8),
         crs=CRS({'init': 'epsg:4326'}),
@@ -388,10 +412,12 @@ def test_geotiff_tiled(default_computation_resources, general_uuid):
         layer_name='Test Raster',
         caption='Raster caption',
         resources=default_computation_resources,
-        filename=str(general_uuid),
+        filename='test_file',
     )
 
-    with rasterio.open(generated_artifact.file_path) as generated_content:
+    with rasterio.open(
+        default_computation_resources.computation_dir / generated_artifact.filename
+    ) as generated_content:
         assert generated_content.profile['tiled']
         assert generated_content.profile['blockxsize'] == 512
         assert generated_content.profile['blockysize'] == 512
@@ -421,7 +447,7 @@ def test_disallow_colormap_for_incompatible_dtype(default_computation_resources)
         )
 
 
-def test_rasterinfo_from_rasterio(default_computation_resources, general_uuid):
+def test_rasterinfo_from_rasterio(default_computation_resources):
     with rasterio.open(f'{os.path.dirname(__file__)}/../../resources/test_raster_a.tiff') as raster:
         # the Colormap type is the correct type for rasterio colormaps
         # noinspection PyTypeChecker