+++ test/base/artifact/test_image.py
@@ -9,8 +9,9 @@ def test_create_concise_image_artifact(default_computation_resources, general_uu
     expected_artifact = _Artifact(
         name='Test Image',
         modality=ArtifactModality.IMAGE,
-        file_path=Path(default_computation_resources.computation_dir / f'{general_uuid}.png'),
+        filename='test_file.png',
         summary='Image caption',
+        correlation_uuid=general_uuid,
     )
     expected_content = Image.new(mode='RGB', size=(2, 2), color=(153, 153, 255))
 
@@ -19,9 +20,11 @@ def test_create_concise_image_artifact(default_computation_resources, general_uu
         title='Test Image',
         caption='Image caption',
         resources=default_computation_resources,
-        filename=str(general_uuid),
+        filename='test_file',
     )
-    generated_content = Image.open(generated_artifact.file_path, mode='r', formats=['PNG'])
+    generated_content = Image.open(
+        default_computation_resources.computation_dir / generated_artifact.filename, mode='r', formats=['PNG']
+    )
 
     assert generated_artifact == expected_artifact
     assert generated_content.convert('RGB') == expected_content
@@ -33,12 +36,13 @@ def test_create_extensive_image_artifact(
     expected_artifact = _Artifact(
         name='Test Image',
         modality=ArtifactModality.IMAGE,
-        file_path=Path(default_computation_resources.computation_dir / f'{general_uuid}.png'),
+        filename='test_file.png',
         summary='Image caption',
         description='Nice graphic',
         sources=default_sources,
         tags=default_association_tags,
         primary=False,
+        correlation_uuid=general_uuid,
     )
     expected_content = Image.new(mode='RGB', size=(2, 2), color=(153, 153, 255))
 
@@ -53,9 +57,11 @@ def test_create_extensive_image_artifact(
         description='Nice graphic',
         sources=Path(__file__).parent.parent.parent / 'resources/minimal.bib',
         tags=default_association_tags,
-        filename=str(general_uuid),
+        filename='test_file',
     )
-    generated_content = Image.open(generated_artifact.file_path, mode='r', formats=['PNG'])
+    generated_content = Image.open(
+        default_computation_resources.computation_dir / generated_artifact.filename, mode='r', formats=['PNG']
+    )
 
     assert expected_content == expected_content_copy, 'Method input should not be mutated during artifact creation'
     assert generated_artifact == expected_artifact